Move get_ksctl_cli to kms.py
